+++ src/builtins/vox.rs
@@ -1,11 +1,12 @@
 use std::env;
-use std::fs::{self, read_dir};
-use std::io::Write;
+use std::fs;
 use std::path::Path;
 
+use crate::builtins::utils::print_stderr_with_capture;
+use crate::builtins::utils::print_stdout_with_capture;
 use crate::parsers;
-use crate::shell;
-use crate::types;
+use crate::shell::{self, Shell};
+use crate::types::{self, CommandResult, CommandLine, Command};
 
 fn in_env() -> bool {
     if let Ok(x) = env::var("VIRTUAL_ENV") {
@@ -29,21 +30,25 @@ fn get_envs_home() -> String {
     home_envs
 }
 
-fn list_envs() -> i32 {
+fn get_all_venvs() -> Result<Vec<String>, String> {
     let home_envs = get_envs_home();
     if home_envs.is_empty() {
-        println_stderr!("you need to set VIRTUALENV_HOME to use vox");
-        return 1;
+        let info = String::from("you need to set VIRTUALENV_HOME to use vox");
+        return Err(info);
     }
     if !Path::new(home_envs.as_str()).exists() {
         match fs::create_dir_all(home_envs.as_str()) {
             Ok(_) => {}
-            Err(e) => println_stderr!("fs create_dir_all failed: {:?}", e),
+            Err(e) => {
+                let info = format!("fs create_dir_all failed: {:?}", e);
+                return Err(info);
+            }
         }
     }
 
+    let mut venvs = Vec::new();
     let pdir = home_envs.clone();
-    if let Ok(list) = read_dir(home_envs) {
+    if let Ok(list) = fs::read_dir(home_envs) {
         for ent in list {
             if let Ok(ent) = ent {
                 let ent_name = ent.file_name();
@@ -52,24 +57,24 @@ fn list_envs() -> i32 {
                     if !Path::new(full_path.as_str()).exists() {
                         continue;
                     }
-                    println!("{}", path);
+                    venvs.push(path);
                 }
             }
         }
     }
-    0
+
+    Ok(venvs)
 }
 
-fn enter_env(sh: &shell::Shell, path: &str) -> i32 {
+fn enter_env(sh: &Shell, path: &str) -> String {
     if in_env() {
-        println_stderr!("vox: already in env");
-        return 1;
+        return format!("vox: already in env");
     }
+
     let home_envs = get_envs_home();
     let full_path = format!("{}/{}/bin/activate", home_envs, path);
     if !Path::new(full_path.as_str()).exists() {
-        println_stderr!("no such env: {}", full_path);
-        return 1;
+        return format!("no such env: {}", full_path);
     }
 
     let path_env = format!("{}/{}", home_envs, path);
@@ -79,22 +84,22 @@ fn enter_env(sh: &shell::Shell, path: &str) -> i32 {
     tokens.push((String::new(), path_new));
     shell::expand_env(sh, &mut tokens);
     env::set_var("PATH", &tokens[0].1);
-    0
+    String::new()
 }
 
-fn exit_env(sh: &shell::Shell) -> i32 {
+fn exit_env(sh: &Shell) -> String {
     if !in_env() {
-        println_stderr!("vox: not in an env");
-        return 0;
+        return String::from("vox: not in an env");
     }
+
     let env_path;
     match env::var("PATH") {
         Ok(x) => env_path = x,
         Err(_) => {
-            println_stderr!("vox: cannot read PATH env");
-            return 1;
+            return String::from("vox: cannot read PATH env");
         }
     }
+
     let mut _tokens: Vec<&str> = env_path.split(':').collect();
     let mut path_virtual_env = String::from("${VIRTUAL_ENV}/bin");
     // shell::extend_env(sh, &mut path_virtual_env);
@@ -109,26 +114,47 @@ fn exit_env(sh: &shell::Shell) -> i32 {
     let env_path_new = _tokens.join(":");
     env::set_var("PATH", &env_path_new);
     env::set_var("VIRTUAL_ENV", "");
-    0
+
+    String::new()
 }
 
-pub fn run(sh: &shell::Shell, tokens: &types::Tokens) -> i32 {
-    let args = parsers::parser_line::tokens_to_args(tokens);
+pub fn run(sh: &mut Shell, cl: &CommandLine, cmd: &Command,
+           capture: bool) -> CommandResult {
+    let mut cr = CommandResult::new();
+    let tokens = cmd.tokens.clone();
+    let args = parsers::parser_line::tokens_to_args(&tokens);
     let len = args.len();
-    if len == 1 {
-        return list_envs();
+    let subcmd = if len > 1 { &args[1] } else { "" };
+
+    if len == 1 || (len == 2 && subcmd == "ls") {
+        match get_all_venvs() {
+            Ok(venvs) => {
+                let info = venvs.join("\n");
+                print_stdout_with_capture(&info, &mut cr, cl, cmd, capture);
+                return cr;
+            }
+            Err(reason) => {
+                print_stderr_with_capture(&reason, &mut cr, cl, cmd, capture);
+                return cr;
+            }
+        }
     }
 
-    let subcmd = &args[1];
-    if len == 2 && subcmd == "ls" {
-        list_envs()
-    } else if len == 3 && subcmd == "enter" {
-        enter_env(sh, args[2].as_str())
+    if len == 3 && subcmd == "enter" {
+        let _err = enter_env(sh, args[2].as_str());
+        if !_err.is_empty() {
+            print_stderr_with_capture(&_err, &mut cr, cl, cmd, capture);
+        }
+        return cr;
     } else if len == 2 && subcmd == "exit" {
-        exit_env(sh)
+        let _err = exit_env(sh);
+        if !_err.is_empty() {
+            print_stderr_with_capture(&_err, &mut cr, cl, cmd, capture);
+        }
+        return cr;
     } else {
-        println_stderr!("vox: invalid command");
-        println_stderr!("usage: vox (ls | enter <env-name> | exit)");
-        1
+        let info = "cicada: vox: invalid option";
+        print_stderr_with_capture(&info, &mut cr, cl, cmd, capture);
+        return cr;
     }
 }